# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for SIMAS (Sistem Informasi Manajemen Sekolah) backend.",
        "title": "SIMAS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Support",
            "email": "support@simas.id"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/sekolah": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List semua sekolah dengan filter status, search, dan pagination. Super Admin only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sekolah"
                ],
                "summary": "List sekolah dengan filter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter status: pengajuan, menunggu_verifikasi, aktif, nonaktif, ditolak",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search nama atau NPSN",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{data[], total, page, limit}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Submit data sekolah tanpa file. Response mengembalikan sekolah_id dan tracking_code. Step 2: upload dokumen via POST /sekolah/:id/dokumen",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sekolah"
                ],
                "summary": "Step 1: Pengajuan baru sekolah (data saja)",
                "parameters": [
                    {
                        "description": "Data sekolah",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SekolahRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "{id, tracking_code, status}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/sekolah/lengkap": {
            "post": {
                "description": "Submit data sekolah + 4 file PDF legalitas via multipart/form-data. Status langsung \"menunggu_verifikasi\". Field data sama dengan SekolahRequest, file: akta_pendirian, nib, sk_pendirian, siop.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sekolah"
                ],
                "summary": "Pengajuan baru sekolah dengan data + dokumen legalitas sekaligus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nama sekolah",
                        "name": "nama",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "sd_mi",
                            "smp_mts",
                            "sma_smk_ma_mak"
                        ],
                        "type": "string",
                        "description": "Jenjang",
                        "name": "jenjang",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email sekolah",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "NPSN",
                        "name": "npsn",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Alamat",
                        "name": "alamat",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Provinsi ID",
                        "name": "provinsi_id",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Kabupaten ID",
                        "name": "kabupaten_id",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Kecamatan ID",
                        "name": "kecamatan_id",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Desa ID",
                        "name": "desa_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Kode pos",
                        "name": "kode_pos",
                        "in": "formData"
                    },
                    {
                        "type": "number",
                        "description": "Latitude",
                        "name": "latitude",
                        "in": "formData"
                    },
                    {
                        "type": "number",
                        "description": "Longitude",
                        "name": "longitude",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Telepon",
                        "name": "telepon",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Website",
                        "name": "website",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Yayasan",
                        "name": "yayasan",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Nama kepala sekolah",
                        "name": "kepala_sekolah_nama",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "NIP kepala sekolah",
                        "name": "kepala_sekolah_nip",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD",
                        "name": "tanggal_berdiri",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Nomor SK pendirian",
                        "name": "sk_pendirian",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Akta Pendirian (PDF, max 3MB)",
                        "name": "akta_pendirian",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "NIB (PDF, max 3MB)",
                        "name": "nib",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "SK Pendirian (PDF, max 3MB)",
                        "name": "sk_pendirian_file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Surat Izin Operasional (PDF, max 3MB)",
                        "name": "siop",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "{id, tracking_code, status, status_label, message}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/sekolah/progress": {
            "get": {
                "description": "Cek status pengajuan sekolah menggunakan tracking_code. Tidak perlu autentikasi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sekolah"
                ],
                "summary": "Track progress pengajuan sekolah (public, no auth)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tracking code dari response pengajuan",
                        "name": "tracking_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SekolahProgressResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.SekolahProgressResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "state_logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SekolahStateLogSummary"
                    }
                },
                "status": {
                    "type": "string"
                },
                "status_label": {
                    "type": "string"
                },
                "tracking_code": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.SekolahStateLogSummary": {
            "type": "object",
            "properties": {
                "catatan": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "status_baru": {
                    "type": "string"
                },
                "status_label": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}